+++ frontend/src/app/app.jsx
@@ -11,7 +11,7 @@ import DocumentTitle from 'react-document-title';
 
 import QAppBar from './QAppBar';
 import QDrawer from './QDrawer/QDrawer';
-import Kids from './Kids';
+import Kids from './Kids/Kids';
 import styles from './styles';
 
 injectTapEventPlugin();
